PB-332: Removing zindex unit-tests
Those tests were testing the deprecated function to compute z-index in the store
that was not anymore used and have been removed.

I tried to unit-test the new z-index composable computation but unit-testing
of such composable is quite complex and I stopped trying to do it after a little
while.