Fix tippy on long layer name in external layer and geo catalogue
On the external layer catalogue as well as the geo catalogue, long entry list
where not truncated and no tippy was added.

Also changed this tippy theme to use the default one and put it on the right.
And fix an issue when the tippy was re-initialize due to a size changes, multipe
tippy were added.