Move "print" job logic and management out of area composable
The print area preview composable was doing the preview, the printing and also coffee. It is now only responsible for the preview area.
I also took the opportunity to remove from the store things that weren't making much sense being there, as no one else as the print menu section was using them (and as the menu section is now the one triggering the print job, it's much clearer this way)