Added e2e tests for tippy on text truncated and try to solve component test flakyness
Make sure that the text truncate tippy works on the active layer list and
on the external map list.

Also try to fix some flakyness on the text truncate component. Somehow it seems
that triggering the tippy not always works using realHover, and that for now
the trigger('mouseenter') seems more stable.