Fix two bugs regarding selected features
The new @features param wasn't checking that it was looping over features with layer, creating error when selecting drawing features
The geometry of a feature came sometimes directly from the store, and was shallow-cloned only by reproject utils. Vuex was then raising an error because reproject attempted to modify something coming from the store. We now deep-clone the geometry before passing it to reproject