PB-361: Avoid to load external WMTS capabilities multiple times
To avoid to load the get capabilities for each layers, we now save the WMTS
options in the layer object.

Due to this we need to register the proj4 projection earlier than in the open
layer map because the get capabilities are parsed during the startup before
openlayer map is loaded.