PB-370: Clear preview layer on selection
The preview layer was not clear when selecting the layer, which let the layer
on top of the stack until a new layer preview was set. Due to this the user
had the impression that it could not changed its active layer configuration
because the preview was still on top.