PB-89: Solved preview debouncing issues
The preview of a search entry can be quite time processing and needed to have some
debouncing to avoid too many store dispatch. Also if the preview was a geojson
layer that needs to be loaded we needed a way to cancel the load if it was
still ongoing when clearing the preview.

To solve those issue the preview layer is delegated to the search list component
which is responsible to send the preview dispatch. This also allow to avoid
sending in a row a clear and set dispatch when changing focus from one entry
to the other.

Now we also have the preview when navigating with the keypboard.

The geojson preview loading is also aborted when preview is cleared.