PB-320: add time slider URL parameter
Add the 'isTimeSliderActive' parameter in the UI store. This parameter will be used to activate the time slider and to store the time slider year between use, in a way similar to the compare slider.
Add the 'timeSlider' URL parameter which will activate the time slider if there is at least one time enabled layer visible.

PB-320: time slider no longer reset 'previewYear'

	as the time slider won't be activated by previewYear, we are no longer resetting previewYear
	also moved youngest year and oldest year to the config, as it's used both in the URL parameter class and the time slider

PB-320: Using isTimeSliderActive in the code

	Used isTimeSliderActive parameter for time slider activation, keeping the previewYear in memory between activations.