PR review
CSS only approach to grid for FeatureList, we will have to wait for CSS masonry to be implemented across all browsers before using it.
The JS approach was a bit unstable when going to Chrome with the fixed width I've set to each category, making them overlap each others. With CSS grid it is much more predictable between browsers (even though we lose the masonry effect when more categories than columns are shown...)