PB-396: Fix search bar focus navigation
When closing the search bar result we focus on the input in order to be able
to directly change the input again. When selecting a result entry we then
close the results but also focus on the search input.

Now the ESC key is a toggling result and when pressing enter on the search
input we jump to the first result (opening the result if needed)