PB-426: Fallback to service search for invalid what3words
Some layers id might be of 3 words separated by dot (e.g. ch.bfe.energieforschung)
which could also be a valid what3words address, therefore we need a fallback
to service search.

Of course now we still the issue if a valid what3words is also a valid layer
ID, the search will always jump to the what3words location, but hopefully this
is quite un-probable.