PB-418: bugfix opacity in WMS layers
Issue : when modyfing the opacity of an external WMS layer and setting it to 0, it would instead revert to its default config value (1 in most cases)

Cause : Opacity of a layer could be set by its parent (parentLayerOpacity). When this was set to 0, the computed check `parentLayerOpacity.value || wmsLayerConfig.value.opacity` would consider the 0 to be a 'false' and use the config value instead.

Fix : We switch the || operator to the ?? operator, which will only use the config value if the parentLayerOpacity is null. We did this change in all layers type to be on the safe side.