PB-422: Fix social network sharing and added linkedin
The social network sharing suffer from a DOM exception. The main issue was
that we used a library that wasn't anymore maintained since 2 years and not
officially supported vue3.

I changed the library by a new one and added linkedin as well.

Note the new library had no support for qrcode.