PB-422: Fix mailto: body in windows outlook
It seems that if the title is empty the mailto: on windows with outlook put
the body inside the subject as follow : `&body=https://s.geo.admin.ch/laksdjflasdf`
instead of putting the url in the email body. Having a title/subject solve this
issue.

On linux the body don't work at all...