PB-438: Fix transparency, time and lang of WMS layers in print
WMS layer based on the WMS specification, prints background in white unless
the query parameter TRANSPARENT is set to true. The WMS parameter were not
passed to print. Due to this external layers were not transparent and layer with
time and/or lang did not print the correct content.

This was a bug in the geoblocks/mapfishprint library, see https://github.com/geoblocks/mapfishprint/pull/30

Geoadmin layers have a bug and their default behavior is transparent=true instead
of false, see https://jira.swisstopo.ch/browse/PB-439

Also fix