PB-378: handle profiles request with too many points
Issue : the profile backend has a query limit of 5000 coordinates. This will be updated in the future, but in the meantime we need to inform users which reach this limit why their query is failing.

Fix : Upon receiving a 413 error from the backend which tells us that the coordinates limit has been reached, we send a profile error with a translation key for the message.