PB-437: Fix print scale
The print scale on mobile was empty. This was due to the fact that we took
the menu width in the computation for the scale.

Now we don't take the menu width for the scale computation and we move the
print center a bit down due to the header overlapping the map.

Also did some renaming to make the code clearer