PB-447: legacy parser not switching to map view
Issue : in some occasions, the router plugins would process the legacy query correctly, but not switch to the map view once it was done, staying stuck in the legacy view forever and not showing anything.

This happened when there was no change to make to the query due to default parameters not being in the URL, but still needing a redirect to the map view.

Fix : If we change the store and the view should change, we ask for a query change too