PB-368 : add "load more features" button in feature list
Also load 50 features when selecting a rectangle on the map (instead of 10)

The piece of code that was triggering the identification of feature was a store plugin, I thought it would make more sense to move all this to the feature store. Especially as I needed to write a second "get more feature" way of loading things. So now the store plugins only triggers a dispatch instead of loading the feature itself.