PB-447: legacy reload parameter in store
We need to be able to send the router to a 'next' route in some rare instances where the legacy parser was called, but there was no default parameter at all.
To do this, we use an app parameter which is set to true by the legacy parser, and set to false the moment the store sync router makes a redirect.

Also adding some debug logs