PB-490: move 3d zoom buttons to right toolbox
Issue : The zoom in and out buttons were added to the cesium viewer near the compass, when it is a bit more logical to have them stay where they are in the 2D viewer.

Fix : We move the logic of those buttons to the toolbox zoom buttons, use that logic when we are in 3d, and removed the cesium buttons.

temp