PB-471 : better handling of drawing floating tooltip
Placing it depending on mobile/desktop drawing elements, and adding checks at tooltip mount to make sure it is within the given limits.

Exporting a bunch of extra SCSS variable and gathering them all in a single JS file to ease their imports down the line.

Adding a bit of padding while editing a feature in the infobox (UI was glued to the border of the screen otherwise...)