PB-501 Rework the data flow in the time slider
The data flow of the time slider was overly complicated and sometimes
lead to triggering setting the same value twice. Tried to remodel the
route that the data takes. Everything now more or less writes to the same
computed property, which triggers setting the store if the data is correct.

In my opinion, with this solution, there are less side effects and (almost)
a single source of truth for the year. Two watchers could be eliminated.