PB-518: Removed time slider on 3D view
Currently on Cesium changing the year is removing/adding the layer which
is not so nice when using the slider play feature as it result to a flickering
effect.

The way Cesium is configured don't allow us to do layer update. It seems that
we could do it with a state management system but would required a big rewrite
of 3D.