PB-496: Fix external layer error and time slider issue
At startup we never know the projection of the external layers, the projection
is taken from the capabilities that needs to be loaded first. Therefore don't
test projection upon creation and don't do any fallback.

Also fixed an issue when a layer has only one timestamp, the time slider was
displayed but the user could not change the timestamp.