Script to generate reports on translation use
The script is used to create three files.

report.json is a json file which contains all translation keys found, state their occurrences across the code, has their translation when there is one available and state if the key is a variable called in the code (in which case, it would be difficult to find all keys that could match this one)

report.txt is a human readable version of the same report

non_explicitly_used_keys.txt is a report which lists all translations for which there was no explicit translation found (i.e : all translations currently specified, but not in the previous reports). We cannot guarantee that they are not used.

We could still upgrade the script by checking for ternary operators which gives explicit values, but this will suffice for a first draft