PB-472 : profile in the context of the new floating tooltip
Changes how the profile handles itself when the floating tooltip is active, closing the profile now only unselect the profile feature (and do not close the selected features anymore)

While the feature info is at the bottom, the profile is now shown as an "off-shoot" or second tab of the selected features, taking the whole width of the screen. Adding a "Back" button at the top of the infobox panel to go back to feature details.

Rework of the profile width management so that it doesn't overflow its parent. Also adding a little spacer between each profile information (going Composition API in the process)