PB-521: fix geolocation tracking and several geolocation improvements
The geolocation tracking on non touch device was canceled as soon as the mouse
moved on the map, even if not action was undertaken ! Now we use the setCenter
to cancel the tracking, unless the setCenter is triggered by the geolocation.

Also did some improvement using the geolocation API by allowing cached location
and setting a timeout.

Also when re-activating the geolocation we first set the previous location
to make the app more reactive.