PB-534: transparency slider works in real time
Issue: The transparency slider did not change the opacity in real time when users were moving the slider, only applying the changes when it was released

Fix: When clicking on the slider, it sets an interval which updates the opacity as long as we're handling the slider. This interval is set to 100ms, which allows to see the changes happen often enough to work, without burdening the backend with too much commits. We also reduce the number of commits by never committing the same value as the one stored.