PB-583 : use backend service to get LV03 <> LV95 conversion right
Using proj4 isn't accurate enough on a geodesic point of view, as the conversion isn't linear. There's a backend service that does the computation in a proper manner, so that's what is being used here when showing a LV03 value in the location popup.
Tracking the mouse position while moving the cursor can still be done with a simple matrix transform, it's "good enough"