PB-449: Add a minimize button to give feedback and report problem
The first idea was to use movable but this cause trouble with the content of
the window not being at a fix size like for the tooltip but dependent on the
window content.

As simpler alternative I added the minimize button which let the user minimize
the window which is for the use case good enough.