PB-546: Fix print overlay not being centered on HiDPI and Retina devices
Somehow, probably due to an openlayer update, the print overlay was not anymore
centered on the map on HiDPI and Retina devices (e.g. Iphone, Dell XPS 15, HP ZBook G9).

After some test it turns out that the issue arised only when the DEVICE_PIXEL_RATIO
was not 1. Removing this ratio from the computed print area solved the issue.
This ratio in the print area has been introduced in 2014 in the old mf-geoadmin3
project (see https://github.com/geoadmin/mf-geoadmin3/pull/1359)
in order to solve the same issue observed here on HiDPI and Retina
device. This issue has been probably solved internally in openlayers now.