PB-546: Improved default print scale computed from map size
Sometimes the initial default print scale was under the header which is not so
nice. This was due to the fact that the map is under the header and that during
the scale computation we took only one header height in consideration but the
print overlay is then center on the whole map, not only on the visible map.
To avoid this issue we take twice the header size now.