Fix uncaught exception when using drag selection
When using CRTL + selection when we already had a drag selection, an uncaught
exception occurred. Although everything seems to work fine, having uncaught exception
is not so good.

So now in order to avoid this exception we first need to remove the selection
layer in every case before re-adding it if needed.