PB-585: make tooltip reactive to language change
Issue : When changing languages, the tooltip did not update to the chosen language and we had to close it and select the features again to see the change
Cause : The features selection process uses calls to the backend that are language dependant. Upon changing the language, we did not change the selected features, meaning they were still in the store with the old language tooltip within.
Fix : A watcher checks language changes when there is a tooltip and update the selected Features

Still to do : currently, changing the language removes the drawings from the selection. Editable Features Selection seems to behave strangely (I can't seem to retrieve them from the store). I'll investigate this further.