PB-585: solving issue with drawings when updating layers
Issue : When we updated the layers, we tried to fetch the drawings from the backend as if they were ordinary layers. This did not work and as a result, the features list had an occurrence where there was a layer with no features selected.
Fix : We add back the initial features if the feature list would be empty after updating.

some linting