PB-412: Allow the print of elevation profile
Elevation profile were not printed because they need to be redrawn using javascript
and could therefore not be added to the print modal window purely as html code.

For this the infobox content has been moved into a vue component so that it can
be reused as slot component in the print button.

Also improved some print css (border and padding) and avoid printing some buttons.