PB-634: Fix feature selection of layer "Moorlandschaften AuLaV"
This layer requires to give the coordinate and other params in order to get
the correct feature.

So in the htmlpopup and in the get feature request we pass the coord, imageDisplay
and mapExtent parameter like in the identify request. This was done like this
in the old viewer. No idea why we need those parameters though.