PB-675: Clean up TimeSlider code
Try to make the code cleaner, more robust and easier to understand (not sure
I achieved this though).

One major changes is that for the "all" hack, now in the url we don't use anymore
@year=9999 but @year=all (legacy timestamp with 9999 are still working and are
translated to @year=all)

Also added support for @year=none when a layer as no data for timeslider
preview year.

I also tried to document the layer attributes and all special cases

I also created a ticket PB-680 to do more clean up concerning the "all" hack
in future.