PB-675: Removed the preview year in layer component
Now the layer timestamp to use is always set to the layer config so there is
no need to look at the preview year of the time slider down the way. This
simplify the code.

Note the preview year is still used as URL param to create permalink with
timeslider activated at a particular year.