PB-669 : separate html popup request from getFeature
identify request already provides us with all what we need (feature data), the only thing missing is the HTML popup.

This means getFeature request is now only called when the app is loaded with pre-selected features.

This reduces the number of calls to our backend for an identify on the map (should be a bit faster)