PB-675: Fix timeSlider reaction when new layer is added/visible
Before when the timeSlider was active and we added a layer, the new layer did
not set its current year to the time slider year until the slider changed.

Also reduced the number of dispatch by only dispatching the set year to layer
that don't have yet the correct year.