PB-420: improve legend css
Issue : when displaying a legend, the styling made it sometimes difficult to read which attributes had which value, as the columns were quite separate from each other.
Fix : When not on mobile, and not in portrait mode, we reduce the width of the table, making the columns closer to each other and making it easier to read.