PB-186: pre processing tool for URL parameters
Issue : some parameters needed to be handled in a special way before we load the map viewer, and this pre-processing happened within the legacy router. This worked fine, but this was not the place to handle it.

Fix : we now have an additional router, called only once between the legacy view and the map view, which handles the pre-processing of parameters.
We ensure this pre processor wait for the legacy parser to finish its work before applying its own changes.
For now, the only parameter with a special process is the 3d, for which we ensure we have the correct spatial reference.