PB-707 : show tooltip on iframe
The old viewer had always shown tooltip when showing the app in embed mode.
With our app we changed that a bit, and that breaks a lot of our customer's integrations (through iframe).

Revert that so that the first feature selection in an iFrame shows a tooltip, only the pre-selection of feature should be impacted by the featureInfo URL param (or showTooltip in legacy)