PB-716 : guess WGS84 coordinate order in the searchbar
without world-wide coverage, it doesn't really make sense for our users that we do not do the same guessing as the old viewer was doing. So adding it back.

This will need to be discussed/re-thinked when we will be adding world-wide coverage, as an inverted WGS84 is still valid in this context (pointing to something outside CH...)