PB-661 : adapting drawing test to new OL singleclick event handler
OL waits 250ms before checking if the click was a single, or double click (or if it moved).
So some tests needed some adaptation, because they fired some (un-retryable) checks right after clicking on the map, without a way of waiting on a backend request to take these 250ms into account.