PB-751: Improved search bar show result behavior
Do not use anymore the on focus event to toggle the results of the search bar.
This was an issue if we wanted to focus the search bar on app startup and that
the swisssearch query param was set, then it would have open the app with the
result open, which was not very nice.

Now I added a dropdown button when the search has result which makes clearer
to the user that there is result even if the results are hidden. The results
are toggle by click event and not anymore by focus event.

Also the ESC button was toggling the results which was not very intuitive and
on the old viewer the ESC button was used to cleared the input, so I reverted
to the old viewer behavior.

Also make sure when the input is cleared that the result is hidden.