PB-747: Fix potential startup race condition
During the startup if the app.isReady was set too early, the initial URL parsing
was not done which would have left the application blank.

We observed such behavior on the prod deployment which was only randomly reproducable.
I suspect that this was the issue but cannot tell for sure.

I also improved logging of the startup procedure.