PB-786: Fix geolocation error management
While using the geolocation you have time to times the `Unknown error` window
that appears. This is due to the fact that the geolocation can return an error,
but  while using the app on the field I noticed that it happens sometimes just
once and then everything seems to work fine. In this case the error windows is
quite annoying as you have to close it and if it happened during the initial
get current position you had to disable and re-enable the geolocation.

So now I improved the error management and do some automatic retries until 3 times
if we are the get current position phase, while in the watcher phase the error
is ignored unless it happens more that 3 times in a row.