PB-395: enhancing highest zoom level
Issue : At the highest zoom level, the mapviewer would not load the highest possible resolution, but the one right before it.
Cause : When we introduced the max resolution to handle client side zoom, we made a slight error in the slicing, and it caused the possible resolutions client side to be smaller (by 1).
Fix : We slice at the correct place.
Bonus Fix : when trying to print at the 1:500 and 1:1000 scales, since the matrixes were sent as parameter by the viewer, we now request the correct tiles.