PB-802: Edit the geoadmin drawing on top
On the old viewer if you would manually import multiple geoadmin drawing (drawing
from our prod backend), going in the drawing menu would copy the drawing
on top of the layer list.

On the old viewer however it would take the top drawing layer even if it is
not visible. However this don't make sense and is disturbing because in this
case if the drawing is invisilbe when opening the drawing mode it is still invisible
until you add the first feature and then the copied feature would sundely appear
making very unintuitive. Therefore now we only looks for visible drawings.