PB-818: Ensuring scales are respected on print
Issue : When we were sending print requests, the scales of the symbols and texts in kmls were off. For icons, they were too small, and the texts were too big.

Fix : We corrected the scales for icons and labels.
Icons used a DPI adaptation of * 90 / 254 which was inherited from the old mapviewer. After some empirical testing, it looks like on the new mapviewer needs a *150 / 254 DPI adaptation
Labels were not using the 'scale' parameter for the fonts at all, and they now also use a 150 / 254 DPI adaptation.

It is important to note that the '254' given here is a computed value, as some layers will have smaller resolutions and their DPI will be used in those situations

Still To Do : Text alignment and ensuring fonts are interpreted correctly on the print side.

PB-818: offsets for the correct sizes

Issue : with the scale being corrected for the icons and the fonts, the offsets were no longer correct.
Thus, we adapt those offsets.

small issue : I had to find ways to set the offsets in an empirical way, to ensure they were place at the correct spot. I'm going to investigate why it works, but for now it works.