PB-804 : don't center on search query when sharing position
Issue : When sharing a position, if there was a search query set, the search query would take priority over the position sharing, removing the shared position pin and centering on the search query result instead.
Fix : When a position is shared, it sets the marker url parameter, which means we can decide that when a marker is present, we should not center on the query result.