PB-804: allow crossHair parameter with only coordinates
Issue : We accepted only two variants of crossHair parameters : crossHair type, and crosshair Type + coordinates. If we tried to set the crossHair to only its coordinates, it wouldn't work.

Fix : When we give coordinates to the crossHair parameter, it uses the 'marker' crosshair Icon by default

corrected an out of bound error