PB-154: Avoid race condition on ol with geolocation tracking
Due to the moveend event that set the center when the user moved the
map by pressing/touching the map and a geolocation arrived during the move operation
the disabled tracking did not work as expected and the map was re-centered
canceling the move action of the user.

So now we are listening to the pointerdrag that happens as soon as we start to
drag and not when dragging is finished. Also moveend happened not only when the
using moved the map, but has soon as the map moved even programmatically.