PB-154: Removed autorotation and geolocation tracking in 3D mode
Re-centering on position in 3D is more complex as in 2d because the of the camera.
If the pitch is not 90° then we need to compute the camera position in order
to have the position centered on the screen. For this moment we don't do this
computation and simply don't re-center the view on geolocation by disabling
geolocation tracking in 3D.