PB-871 : re-arranging app config into multiple files
Trying to gather common config together in a meaningfully named config file.

Adding function to get base URLs, instead of direct access to constants. So that we have the opportunity to override base URLs. I didn't want to go through the store to do that. Doing so would create a lot of circular dependencies between the store, the config and components.
As it's a debug tool, that's "good enough" this way.