PB-984: add error feedback for all URL parameters
Issue : when modifying the URL directly, users received no feedback about what they did wrong when making mistakes, especially on a productive environment where most logs are deactivated.
Fix : we now give a feedback for every parameter that is wrongfully filled. We also added some extra cases for the layers, where we want to give a feedback on each individual layer, and for the time slider when everything is in order, but there are no layers supporting it.