PB-896 : show un-simplified GPX track
but still simplify the track before sending it to the profile backend.
This should reduce the coastal paradox we had before when sending raw GPX track, while keeping user "happy" as they will se their raw track on the map

Using TurfJS to do the simplification, as it is also using a Ramer-Douglas-Peucker algorithm (same as OpenLayers) and it would be more difficult / add more complexity to the code to reparse all features as OL features there.