PB-960: handle long press events error on some mobiles
Issue : on some phones, a long press event would also fire a single press event afterwards, causing the application to open both a location popup and an infobox.
Cause : setting the longClickTriggered boolean was not registered while in the longClickTimeout function, most likely due to a scoping issue.

Fix : we set the longClickTriggered boolean in the 'onMapRightClick' event instead. We also added a small parameter in the function to tell if we need to set up the flag or not, as desktop users won't need this.

PB-960: add updateLongClickTriggered to event

- We add the updateLongClickTriggered to the 'event' item in onMapRightClick