PB-827: Add warning when using http resources
Issue : Supporting http resources with the mapviewer is being rolled out by late January at the latest. We need to inform people still using them to make the change

Fix : We add a feedback warning specifically saying that there are resources in http, and that the user should change that.

remove excess log