PB-722 : improve GeoJSON, KML and GPX support in 3D
Rework of the GeoJSON style converter, so that it can deal with triangle/square (any non-circle) shapes by using the OL icon generator we already have in the code.

Adding support for opacity for all 3 types of layers (GeoJSON, KML and GPX)

Clamping internal KMLs (from the drawing module) to the ground for better readability. keeping external non-clamped, so that they may give use the height (z coordinate). We do not use this z axis on drawing, so it makes sense that we clamp our stuff to the ground.