PB-984: quick fix for local layers
Issue: when reloading the page, local layers trigger the error feedback, telling them they made a mistake, when technically they did not do anything wrong.

Fix : we assume local layers errors are due to actualisations / links sharing, and that we should warn the users they're trying to get a local file through the URL, which doesn't work, and gives them a feedback about it in the form of a warning.