PB-984: Not sending warnings when not necessary
Issue : in some cases, warnings or errors could be sent when unnecessary, as the router would loop more than once when mutations occur, and thus go through the URL query watcher. This was most visible with local external layers, which would be checked when not needed (for example, when switching from 2d to 3d).

Fix : If the value in the store is the same as the one in the query, we no longer validate the input.