PB-1028 : add a parser for each type of file we want to support
Instead of having a big chunk of code that does the heavy lifting (loosely copied or similar in places like the load-kml and load-gpx plugin, plus the drag and drop handler) I gathered all this logic in one place, giving it the necessary options/config to handle all situation and be written only once.

There is now only one way of dealing with a KML/KMZ/GPX file from external/local source, the dedicated parser.

List of other things done while working on this :
 - changing updateLayer action to receive a layer ID instead of an index (was never used because of the hassle to get this index). I replaced uses of updateLayers when it was used for a single layer.
 - move all extent utils in... extentUtils.js
 - FileInput can now receive params for i18n placeholders in the error message
 - Fixing a missing rename with the local file disclaimer, it was not the correct tooltip shown for local files because of that
 - Adding a zoom to extent button in the layer list, if the layer provides an extent (meaning KML/GPX and soon COG)