PB-901 : disable drawing name input when no feature are drawn
Adding all relevant i18n keys for this new input.

Showing a tooltip (while disabling the input) telling the user it's not possible to change the drawing name while there is nothing drawn.
This was the easiest way of dealing with this input, as storing the value before a KML was present, or creating an empty one, was really complicated to manage down the line.

So by waiting for a KML (with at least a feature) to be present, we can safely change the KML name without having too much defensive code, or having to store the drawing name in Vuex (or move this part to the DrawingModule, changing a ton of things)