PB-875 : fix potential flakiness of tests
the box selection was happening from the center of the view, and the KML feature that was tested is placed exactly at the center.
That lead to tests failing sometimes on the CI.

So I've changed the box-making function to create a box around the center of the view instead.