PB-1140: add a minimal scale for print specs
Issue : When sending a print spec, if the scale was too small, it would be sent to the backend in a scientific notation (for example : '1.00045e-11px'), which couldn't be interpreted by the backend.

Fix : We set a new minimal scale value to ensure we always send readable specs, which is enforced as long as the scale is greater than 0.