PB-1028 : fix COG loading after rebase error
While working on the latest PR for PB-1028 I made some rebase mistake and didn't detect them before merging.

COG files were loaded entirely once again... So here's the fix, with some additional benefits.

I tried to simplify the FileAPI functions, the ones that load files (either through a simple GET or through service-proxy)
I wasn't possible, with the addition of COG to the mix, to keep FileParser to decide to load content on its own, so I created a function that get the MIME type (and other relevant info on the file) so that each time we need to load a file, we can decide which way (of loading) we want to go (proxy or not proxy)