PB-1121: using pre compose event to handle COG slider
Issue : When importing or toggling visibility on WebGL layers, the layerconfig was updated before the map, and the new layer would not have the prerender and postrender events set.

Fix : OpenLayers Map fire precompose events when webGL layers are involved. This allows us to add a precompose event on the map to add the prerender events just before the first rendering is supposed to happen.