PB-1151: allow partial profiles
Issue : When we request a profile that is partially out of bounds, the inbound parts are not requested. This is caused by running into an error when we parse the out of bounds part of the backend response

Fix : We handle the out of bounds part of the backend response and its special structure.