PB-995 : Cesium WMTS as composable
transform the Cesium WMTS component to composition API

Adding a composable that can be used for WMS layers when we want to move them to composition API too.

Issues fixed by this commit :
- opacity when a layer as a specific config for 3D (like SWISSIMAGE) by giving the 3D config the same opacity as the 2D equivalent.
The opacity of the 3D config was used, but the UI is editing the opacity of the 2D layer.
- error that popped up in the console when switching from 3D to 2D, the camera position is set to null by the unmounted CesiumMap hook and then the sync camera plugin uses the camera value without checking its validity. Adding a little check to remove the error
- External WMTS style wasn't properly applied to the 3D viewer, a good example is to load swissdatacube WMTS with layer diff_all_CH, as its default style isn't named "default" but something else